// apps/web/src/lib/types.ts
export const BiomarkerSchema = z.object({
  id: z.number().int().positive(),
  name: z.string(),
  elab_code: z.string().nullable(),
  slug: z.string().nullable(),
  lab_prices: z
    .record(z.string(), z.number().int().nonnegative())
    .default({}),
});
// ...
export const BiomarkerSearchResponseSchema = z.object({
  results: z.array(BiomarkerSchema.pick({ name: true, elab_code: true, slug: true })),
});